Add buildTree(), filterDuplicates(), and print()
In this commit, I implemented the function to build a tree
and added a function to print the contents of the tree. this
function is courtesy of The Odin Project. All the user needs
to do is call BinarySearchTree.print() and it'll help them
visualize the tree.

I had deleted the mergeSort function because I thought I would
not need it, but when I was implementing buildTree(), I noticed
the tree that was being built had a weird construct. So, I
re-implemented mergeSort() again and used it to sort out an
array that is filtered with filterDuplicates() to not insert
duplicates into the tree. That filtered output is then sorted and
fed to buildTree().

All this process is initialized by calling init() within the
binary search tree object. This is not exposed to the user at all,
as they do not need to know these details.